+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.6.573](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.572...@codecademy/gamut-kit@0.6.573) (2025-12-10)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+### [0.6.572](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.571...@codecademy/gamut-kit@0.6.572) (2025-12-08)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
 ### [0.6.571](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.570...@codecademy/gamut-kit@0.6.571) (2025-12-05)
 
 **Note:** Version bump only for package @codecademy/gamut-kit